refactor(Sentry): Discrete files for sentry kmp wrappers
