+++ SwissTransferCore/Sentry/KotlinThrowableWrapper.swift
@@ -1,0 +1,41 @@
+/*
+ Infomaniak SwissTransfer - iOS App
+ Copyright (C) 2025 Infomaniak Network SA
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
+
+ You should have received a copy of the GNU General Public License
+ along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import Foundation
+import Sentry
+import STCore
+
+public final class KotlinThrowableWrapper: NSError, @unchecked Sendable {
+    public init(kotlinThrowable: KotlinThrowable) {
+        let errorMessage = kotlinThrowable.message ?? ""
+        let userInfo: [String: Any]
+        if let cause = kotlinThrowable.cause {
+            let underlyingError = [KotlinThrowableWrapper(kotlinThrowable: cause)]
+            userInfo = [NSLocalizedDescriptionKey: errorMessage, NSUnderlyingErrorKey: underlyingError]
+        } else {
+            userInfo = [NSLocalizedDescriptionKey: errorMessage]
+        }
+
+        super.init(domain: "kmp.infomaniak.com", code: -666, userInfo: userInfo)
+    }
+
+    @available(*, unavailable)
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
